chore(folio): move comment helpers to commentsHelpers module
`DocxEditor.tsx` carried ~130 lines of pure comment-related helpers
between the type block and the component: `createComment`,
`applyCommentMarkRange`, `removePendingCommentMarkRange`,
`getCommentAuthorKey`, `findSelectionYPosition`,
`getFallbackCommentYPosition`, `getCommentParentId`, plus the
constants `PENDING_COMMENT_ID`, `EMPTY_ANCHOR_POSITIONS`, and the
module-level `nextCommentId` counter. None of these depend on
component state — they take everything they need as arguments. Pull
them out into a sibling `commentsHelpers.ts` module.

While here:

- Wrap the mutable `nextCommentId` counter behind an explicit
  `allocateCommentId()` accessor instead of the bare `nextCommentId++`
  expression at the use site, so the in-process id allocator is a
  named operation.
- Promote the magic `80` fallback offset in
  `getFallbackCommentYPosition` to a named `FALLBACK_TOP_OFFSET`
  constant (also clears up that the second `80` was the same value).

No behavior change. State and the 56 setter callsites stay in
`DocxEditor.tsx`; the hook-based state extraction is a follow-up.
